requirements.txt
Django==1.10.2
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
PyJWT==1.4.2
wheel==0.24.0
